Fix the line deliminator in csv
currently the csv output is `\r\r\n` where the line terminator is `\r\n`
and the extra `\r` is probably being inserted from writing a dict to
csv. I get around this by changing the line terminator to `\n` to output
`\r\n` which will make it compatible with most spreadsheet apps.