fixes osx bug where user couldn't change proxy or connect to network
by default osx passes a -psn arg when launching an app from command
line. this is a process serial number argument and py2app when used with
emulate_argv=True removes it.
commit d2097c15d9eb5ddb12a02dbe4d837b97f091f808 removed this, thus
introducing the -psn arg on launch. this is misinterpreted by the arg
parser as a proxy setting, breaking networking.