Edit install wizard msg to reflect linguistic trends
In the install wizard message, the prompt uses the word "fulfil". While this is an acceptable spelling of the word, Google ngrams indicates that it is twice as common to spell the word with two l's at the end, aka "fulfill", in modern writing. This PR proposes to use the more common spelling.

https://books.google.com/ngrams/graph?content=fulfill%2C+fulfil&year_start=1800&year_end=2000&corpus=15&smoothing=3&share=&direct_url=t1%3B%2Cfulfill%3B%2Cc0%3B.t1%3B%2Cfulfil%3B%2Cc0