kivy: after deleting wallet, try to open saved wallet
saved wallet = "wallet that was last open during a graceful shutdown"

previously we would just try to open "default_wallet". now, if the deleted wallet is not the saved wallet, we open the saved wallet; else try to open "default_wallet". if "default_wallet" does not exist, as before, the wizard launches.